% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_flows.R
\name{clean_flows}
\alias{clean_flows}
\title{Clean Flow Record}
\usage{
clean_flows(discharge, replace_0 = 0.001, replace_neg = NA)
}
\arguments{
\item{discharge}{numeric vector of discharges.}

\item{replace_0}{numeric value or \code{NA} to replace zeros with. Defaults to 0.001.}

\item{replace_neg}{numeric value or \code{NA} to replace negative values with. Defaults to \code{NA}.}
}
\value{
numerioc vector same length as values provided in \code{discharge}.
}
\description{
Function to replace zeros in the flow record with specified value and replace negative discharge values with \code{NA}.
}
