% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_sep.R
\name{lyne_hollick}
\alias{lyne_hollick}
\title{Lyne and Hollick Equation}
\usage{
lyne_hollick(Q, a)
}
\arguments{
\item{Q}{numeric vector of discharge values}

\item{a}{alpha value between \code{[0-1]}}
}
\value{
vector of estimated baseflow values
}
\description{
Filter equation.
}
\keyword{internal}
