% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_sep.R
\name{run_filter}
\alias{run_filter}
\title{Iterate Lyne-Hollick Filter}
\usage{
run_filter(Q, a, n)
}
\arguments{
\item{Q}{vector of discharge values}

\item{a}{alpha parameter}

\item{n}{number of times to itterate filter.}
}
\value{
numeric vector
}
\description{
Simply iterates n runs of the Lyne-Hollick filter.
}
\keyword{internal}
